Adapt include paths for api/Foebud.php
The custom "API" file (which isn't really called through the API
here...) lives in the extension module nowadays; so adapt include
paths to reference the new location directly, rather than the old
location in the main civicrm module -- which was only a symlink now
anyways, and not needed here at all.

Also adapt README.txt to reflect the fact that we don't need the symlink
anymore.