Turn into extension
This is a more or less minimal set of changes to turn the Drupal module
into a "native" CiviCRM extension.

The extension code is created with the Civix code generator.
Specifically, donationreceipts.civix.php is entirely generated;
donationreceipts.php is mostly generated, except for our hook moved here
from the original .module file; and info.xml is based on a generated
template, with the actual data mostly imported from the original .info
file.

This isn't yet any less (or more) hacky than it was as a module -- i.e.
it still needs quite a lot of work...