Drop API sham
The "API" function for generating donation receipts was never actually
invoked through the API (rather, the file is just included directly) --
and indeed there is no reason whatsoever to involve the API machinery
here, as this is never called from external code. So stop playing
make-believe: it was only confusing.

(Leaving the API-like handling of function parameters in place though:
it doesn't make much of a difference either way -- so don't bother
changing it...)