Use name-based heuristics to decide contribution category
Rather than hardcoding the `contribution_type_id` for membership dues
vs. donations, try to guess it from the name. This way it doesn't need
to be adapted for every installation.

It isn't a very robust approach, but probably good enough in practice --
and several orders of magnitude simpler than every other variant I can
think of...