Add menu entry for Jahresbescheinigungen
Instead of the hack we had so far linking it from the contact tab.

Note: the entry is addded dynamically at the end of the "Contributions"
submenu (wherever it happen to reside) with a hook. This is somewhat
problematic, as it can't be moved or otherwise edited with the menu
editor... It will always go back to the automatic entry when the Menu is
rebuilt. (Including after any edits with the menu editor...)

It might be more elegant to add a static menu entry upon install.
However, I haven't found anything in the API providing that
functionality...