Manage receipt templates through CiviCRM UI
When generating receipts, fetch the templates from the
`civicrm_msg_template` DB table, instead of the file system. (After
putting the default templates we ship with the extension into the DB
during installation.)

Rather than implementing our own UI for editing the templates, we are
(mis)using the System Workflow Messages management screen for this.

As the system workflow templates are normally used only for emails, the
management screen is only accessible when CiviMail is active. Also, we
need to add some extra workflow-related DB entries, to keep the database
consistent and to make our templates easily accessible. Still,
`civicrm_msg_template` really seems the most obvious place to store the
templates; and apart from the minor quirks mentioned above, this
approach should be unproblematic.