Default templates: Explicitly position bottom section at page bottom
Use 'postion: absolute' CSS property to make sure that the section
supposed to be at the bottom of the page actually is positioned exactly
there, rather than relying on the rest of the layout fitting just right.

This is much more robust in general. Most notably, it makes the receipts
somewhat independent of the used page format. (As long as it's more or
less A4 with reasonable margins...)

In fact this makes the templates work again out of the box... (Not
overflowing the bottom remark to another page.) This was probably broken
by some dompdf change -- which just proves the point about robustness
:-)