Use CRM_Utils_PDF_Utils::html2pdf()
Instead of invoking dompdf directly, use the wrapper provided by
CiviCRM.

This is more elegant and consistent. For one, wkhtmltopdf (snappy) will
be used instead of dompdf when available.

As html2pdf() sets up the HTML container and head itself (and drops any
supplied in the input...), the templates need adjusting to provide only
the body contents.

Note that now the page format is no longer hard-coded to dompdf's "A4"
format: using the standard PDF format configured in CiviCRM instead --
so the receipts will be rendered with different page margins after this
change...