Clean up all HTML pages and related code
Various improvements to HTML correctness/style; message wording; and
layout of the various output pages. This includes some minor functional
changes in handling of errors and other special cases.

As part of the rewording, make sure to use real umlauts in all ordinary
output, i.e. template files. (But not in error messages or internal
names in php files!)

Note: Some of the template changes allowed for simplifications/cleanups
on the php side of the template handling too.