Default template: More closely reproduce layout of official forms
The previous layout was somewhat arbitrary. Rather try to stick to the
official layout as close as possible, so there is no question why it
looks this way and not another. Organisations can still customise it
according to their wishes of course.

As for actual changes: most notably, don't put the issuer and donor
addresses side by side anymore. (I guess the side-by-side layout has
been deemed convenient for organisations not using a cover letter when
sending the receipts. However, it's not even clear that it was actually
valid if going stictly by the guidelines, as it kinda changed the order
of fields. At any rate, it was much more complex; and it was very much
*not* like the official form... This doesn't seem justified for the
default.)

Also, there is less variation in font sizes and margins now. All in all,
this adjusted layout is simpler, and IMHO also clearer compared to what
we had before.