Makefile: Explicitly store base directory in .zip
When a zipfile is created with zip -r, the directory is always stored
explicitly along with the files it contains. As we are not using zip -r,
the base directory was not stored in our case though. While I'm not sure
this is actually an error (unzip at least doesn't mind), it confused the
CiviCRM .zip handler, preventing automatic installation from the
extension directory through the "manage extensions" UI.