Use new name for Message Templates BAO on CiviCRM >= 4.4
The BAO got renamed from MessageTemplates to MessageTemplate -- so we
need to call the right one depending on the CiviCRM version in use.

This is a minimal change. 4.4 also introduced an API for Message
Templates, which we were missing before -- so ideally we would
transition to using this API now instead of the BAO, in the hope of
avoiding such compatibibility problems in the future... However, that
would be a much more involved change -- especially if we want to
preserve compatibility with the older versions too.